requirements.txt
# ASHES (Autonomous Scientific Hypothesis Evolution System) Dependencies

# Core Framework
langchain==0.1.20
langgraph==0.0.55
openai==1.12.0
anthropic==0.21.3
together==0.2.7

# Vector Databases & Storage
pinecone-client==3.2.2
weaviate-client==4.5.4
chromadb==0.4.24
redis==5.0.3

# Machine Learning & AI
tensorflow==2.16.1
torch==2.2.1
transformers==4.38.2
sentence-transformers==2.7.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1

# Web Framework & APIs
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
aiohttp==3.9.3
websockets==12.0

# Task Queue & Messaging
celery==5.3.6
kafka-python==2.0.2

# Databases
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
neo4j==5.18.0
influxdb-client==1.42.0

# Object Storage
minio==7.2.5
boto3==1.34.69

# Security & Authentication
cryptography==42.0.5
bcrypt==4.1.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Monitoring & Observability
prometheus-client==0.20.0
structlog==24.1.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-jaeger==1.21.0

# Laboratory Automation
pymeasure==0.13.1
pyvisa==1.14.1
pyserial==3.5
opencv-python==4.9.0.80

# Scientific Computing
scipy==1.12.0
sympy==1.12
matplotlib==3.8.3
plotly==5.19.0

# Data Validation & Models
pydantic==2.6.3
pydantic-settings==2.2.1

# Testing & Development
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==5.0.0
black==24.3.0
flake8==7.0.0
mypy==1.9.0

# Configuration & Utilities
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
typer==0.9.0
